requirement.txt
absl-py==1.3.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
boto3==1.34.108
botocore==1.34.108
Brlapi==0.8.3
cachetools==5.2.0
certifi==2024.2.2
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dataclasses-json==0.6.6
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.3.0
docopt==0.6.2
duplicity==0.8.21
ecdsa==0.19.0
email_validator==2.1.1
entrypoints==0.4
exceptiongroup==1.2.1
executing==1.1.1
faiss-cpu==1.9.0.post1
fastapi==0.111.0
fastapi-cli==0.0.3
fastapi-limiter==0.1.6
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.14.0
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2024.5.0
future==0.18.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.19.0
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.8.3
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h5py==3.7.0
httpcore==1.0.5
httpie==2.6.0
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.0
idna==3.7
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.11.0
keyring==23.5.0
kiwisolver==1.4.4
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-google-genai==2.0.5
langchain-text-splitters==0.3.2
langsmith==0.1.143
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.61.0
mistune==2.0.4
mongo==0.2.0
monotonic==1.6
more-itertools==8.10.0
motor==3.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.3
nltk==3.8.1
notebook==6.4.12
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
orjson==3.10.3
outcome==1.2.0
packaging==23.2
pandas==1.5.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pinecone==4.0.0
pinecone-client==4.1.0
pipreqs==0.4.13
platformdirs==3.5.0
pluggy==1.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
proto-plus==1.24.0
protobuf==4.25.5
psutil==5.9.2
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.8.1b0
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.7.3
Pympler==1.0.1
PyMuPDF==1.24.14
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.3.3
pytest-asyncio==0.24.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
redis==5.2.0
regex==2024.5.15
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.3.4
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
selenium==4.5.0
Send2Trash==1.8.0
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.30
stack-data==0.5.1
starlette==0.37.2
sympy==1.12
systemd-python==234
tenacity==8.3.0
termcolor==2.1.1
terminado==0.16.0
threadpoolctl==3.5.0
tinycss2==1.1.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.1.0
toolz==0.12.0
tornado==6.2
tqdm==4.66.4
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
triton==2.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.14
usb-creator==0.3.7
uvicorn==0.29.0
uvloop==0.19.0
validators==0.20.0
virtualenv==20.23.0
wadllib==1.3.6
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.5
webdriver-manager==3.8.3
webencodings==0.5.1
websockets==12.0
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarg==0.1.9
yarl==1.9.4
zipp==1.0.0
